pyproject.toml
[tool.poetry]
name = "test_task_django"
version = "0.1.0"
description = ""
authors = ["suband74 <suband@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
django-mptt = "^0.13.4"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
